added status, changed viewOrders id to order-details

+++ src/scripts/components/events/domEvents.js
@@ -20,22 +20,18 @@ const domEvents = (uid) => {
         orderDetails(orderId);
       });
     }
-    // if (e.target.id.includes('delete-order')) {
-    //   // eslint-disable-next-line no-alert
-    //   if (window.confirm('Want to delete?')) {
-    //     const [, firebaseKey] = e.target.id.split('--');
-    //     deleteOrder(firebaseKey).then((orderArray) => showOrders(orderArray));
-    //   }
-    // }
     if (e.target.id.includes('delete-order')) {
-      const [, firebaseKey] = e.target.id.split('--');
-      deleteOrder(firebaseKey).then(() => {
-        getOrders(uid).then((orderArray) => {
-          showOrders(orderArray);
+      // eslint-disable-next-line no-alert
+      if (window.confirm('Want to delete?')) {
+        const [, firebaseKey] = e.target.id.split('--');
+        deleteOrder(firebaseKey).then(() => {
+          getOrders(uid).then((orderArray) => {
+            showOrders(orderArray);
+          });
         });
-      });
+      }
     }
-    if (e.target.id.includes('viewOrders')) {
+    if (e.target.id.includes('order-details')) {
       getOrders(uid).then((orderArray) => {
         showOrders(orderArray);
       });
@@ -57,7 +53,7 @@ const domEvents = (uid) => {
     }
     if (e.target.id.includes('order-details')) {
       const [, firebaseKey] = e.target.id.split('--');
-      getSingleOrder(firebaseKey).then((orderObj) => orderDetails(orderObj));
+      getSingleOrder(firebaseKey).then((orderObj) => orderDetails(orderObj.firebaseKey));
     }
   });
 };
